webui/requirements.txt
GitPython
Pillow
accelerate

basicsr
blendmodes
clean-fid
einops
gfpgan
gradio==3.32.0
inflection
jsonmerge
kornia
lark
numpy
omegaconf
open-clip-torch

piexif
psutil
pytorch_lightning
realesrgan
requests
resize-right

safetensors
scikit-image>=0.19
timm
tomesd
torch
torchdiffeq
torchsde
transformers==4.25.1
django 